pyproject.toml
[tool.poetry]
name = "jsonipy"
version = "0.1.0"
description = ""
authors = ["Jackie Tung <jackie.tung@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
jsonipy = "jsonipy.main:main"

[tool.poetry.dependencies]
python = ">=3.10"
pyyaml = "^6.0.1"
pyjson5 = "^1.6.6"
python-hcl2 = "^4.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
